solving issue on undefined
The first attempt to allow the user to provide a custom style object led to an unexpected <undefined> error. Turns out the problem had to do with having to populate the object with default values (in case it was not well defined by the user. This is achieved by the function populateStyleObject. Everything works fine now.